# feat: implement automatic geo-detection and performance data persistence