--- a/Cargo.toml
+++ b/Cargo.toml
@@ -67,7 +67,11 @@ sha2 = "0.10"
 blake3 = "1.0"
 
 # Configuration
+figment = { version = "0.10", features = ["toml", "env"] }
 config = "0.15"
+
+# File watching for hot reload
+notify = "8.0"
 
 # Compression
 flate2 = "1.0"
@@ -83,6 +87,10 @@ metrics = "0.24"
 
 # TOML parsing
 toml = "0.8"
+
+
+# Human-readable time parsing
+humantime-serde = "1.1"
 
 # Directory utilities
 dirs = "6.0"