--- a/src/sources/cloudflare.rs
+++ b/src/sources/cloudflare.rs
@@ -209,7 +209,7 @@ impl DownloadSource for CloudflareSource {
     }
 
     fn priority(&self) -> u8 {
-        self.config.priority
+        self.config.priority.min(255) as u8
     }
 
     fn is_enabled(&self) -> bool {