--- a/src/sources/fastly.rs
+++ b/src/sources/fastly.rs
@@ -71,7 +71,7 @@ impl DownloadSource for FastlySource {
     }
 
     fn priority(&self) -> u8 {
-        self.config.priority
+        self.config.priority.min(255) as u8
     }
 
     fn is_enabled(&self) -> bool {