--- a/src/sources/github.rs
+++ b/src/sources/github.rs
@@ -235,7 +235,7 @@ impl DownloadSource for GitHubSource {
     }
 
     fn priority(&self) -> u8 {
-        self.config.priority
+        self.config.priority.min(255) as u8
     }
 
     fn is_enabled(&self) -> bool {